GraphQl : save country_code in setPaymentMethodOnCart, rename setCheckoutTokenGraphQL to SetCheckoutTokenGraphQL
