+++ etc/graphql/di.xml
@@ -2,6 +2,6 @@
 
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
     <type name="Magento\QuoteGraphQl\Model\Cart\SetPaymentMethodOnCart">
-        <plugin name="affirm_checkout_token" type="Astound\Affirm\Plugin\setCheckoutTokenGraphQL" />
+        <plugin name="affirm_checkout_token" type="Astound\Affirm\Plugin\SetCheckoutTokenGraphQL" />
     </type>
 </config>
